+++ example/web/pwa.g.dart
@@ -1,8 +1,0 @@
-import 'package:pwa/worker.dart';
-import 'package:pwa_example/pwa/worker.dart' as custom;
-
-/// Starts the PWA in the worker scope.
-void main() {
-  PwaWorker worker = custom.createWorker();
-  worker.run();
-}