+++ examples/custom_routes/lib/pwa/worker.dart
@@ -3,11 +3,11 @@ import 'offline_urls.g.dart';
 
 /// Creates the PWA worker.
 PwaWorker createWorker() {
-  DynamicCache youtubeThumbnails =
-      new DynamicCache('youtube', maxEntries: 10, noNetworkCaching: true);
-
   PwaWorker worker = new PwaWorker()
     ..offlineUrls = offlineUrls;
+
+  DynamicCache youtubeThumbnails =
+      new DynamicCache('youtube', maxEntries: 10, noNetworkCaching: true);
 
   worker.router.get('https://i.ytimg.com/vi/', youtubeThumbnails.cacheFirst);
   return worker;